requirements.txt
streamlit
pandas
boto3
python-dotenv
plotly
textblob
requests
pyarrow
